requirements.txt
# Core dependencies
requests>=2.31.0
python-dotenv>=1.0.0
selenium>=4.18.1
beautifulsoup4>=4.12.3
webdriver-manager>=4.0.1
google-maps-places==0.1.25
# Development dependencies
pylint>=3.1.0
mypy>=1.8.0 


